General JS improvements
