+++ client/blocks/upe/upe-deferred-intent-creation/payment-processor.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-import { getPaymentMethods } from '@woocommerce/blocks-registry';
 import {
 	PaymentElement,
 	useElements,
@@ -12,10 +11,12 @@ import { useEffect, useState, useRef } from 'react';
 /**
  * Internal dependencies
  */
-import { usePaymentCompleteHandler, usePaymentFailHandler } from '../hooks';
+import {
+	useCheckoutSuccessHandler,
+	useCheckoutFailHandler,
+	usePaymentSetupHandler,
+} from '../hooks';
 import BlikCodeElement from './blik-code-element';
-import { __ } from '@wordpress/i18n';
-import { select } from '@wordpress/data';
 import {
 	getBlocksConfiguration,
 	getStripeElementOptions,
@@ -25,7 +26,6 @@ import {
 	maybeShowCashAppLimitNotice,
 	removeCashAppLimitNotice,
 } from 'wcstripe/stripe-utils/cash-app-limit-notice-handler';
-import { validateBlikCode } from 'wcstripe/stripe-utils';
 import {
 	PAYMENT_METHOD_BLIK,
 	PAYMENT_METHOD_CASHAPP,
@@ -65,7 +65,6 @@ export function validateElements( elements ) {
  * @param {Array}       args.upeMethods          The UPE methods.
  * @param {string}      args.errorMessage        The error message to display.
  * @param {boolean}     args.shouldSavePayment   Whether or not to save the payment method.
- * @param {string}      args.fingerprint         The fingerprint.
  * @param {Object}      args.billing             The checkout billing data.
  *
  * @return {JSX.Element} Rendered payment processor.
@@ -82,7 +81,6 @@ const PaymentProcessor = ( {
 	upeMethods,
 	errorMessage,
 	shouldSavePayment,
-	fingerprint,
 	billing,
 	onLoadError = noop,
 } ) => {
@@ -92,9 +90,6 @@ const PaymentProcessor = ( {
 		useState( null );
 	const [ isPaymentElementComplete, setIsPaymentElementComplete ] =
 		useState( false );
-	const paymentMethodsConfig = getBlocksConfiguration()?.paymentMethodsConfig;
-	const gatewayConfig = getPaymentMethods()[ upeMethods[ paymentMethodId ] ];
-	const isBlikSelected = selectedPaymentMethodType === PAYMENT_METHOD_BLIK;
 
 	// Make sure shouldSavePayment is set to true if the cart contains a subscription.
 	// shouldSavePayment might be set to false because the cart contains a subscription and so the save checkbox isn't shown.
@@ -108,164 +103,6 @@ const PaymentProcessor = ( {
 		hasLoadErrorRef.current = true;
 		onLoadError( event );
 	};
-
-	useEffect(
-		() =>
-			onPaymentSetup( () => {
-				async function handlePaymentProcessing() {
-					if (
-						upeMethods[ paymentMethodId ] !== activePaymentMethod
-					) {
-						return;
-					}
-
-					if ( hasLoadErrorRef.current ) {
-						return {
-							type: 'error',
-							message: __(
-								'Invalid or missing payment details. Please ensure the provided payment method is correctly entered.',
-								'woocommerce-gateway-stripe'
-							),
-						};
-					}
-
-					const { validationStore } = window.wc?.wcBlocksData ?? {};
-					if ( validationStore ) {
-						const store = select( validationStore );
-						const hasValidationErrors = store.hasValidationErrors();
-						// Return if there is a validation error on the checkout fields.
-						if ( hasValidationErrors ) {
-							return;
-						}
-					}
-
-					// BLIK is a special case which is not handled through the Stripe element.
-					if ( ! ( isPaymentElementComplete || isBlikSelected ) ) {
-						return {
-							type: 'error',
-							message: __(
-								'Your payment information is incomplete.',
-								'woocommerce-gateway-stripe'
-							),
-						};
-					}
-
-					if ( errorMessage ) {
-						return {
-							type: 'error',
-							message: errorMessage,
-						};
-					}
-
-					// Check if user tried to save a method that isn’t reusable.
-					if (
-						gatewayConfig.supports.showSaveOption &&
-						shouldSavePayment &&
-						! paymentMethodsConfig[ paymentMethodId ].isReusable
-					) {
-						return {
-							type: 'error',
-							message:
-								'This payment method cannot be saved for future use.',
-						};
-					}
-
-					if ( isBlikSelected ) {
-						validateBlikCode();
-					} else {
-						await validateElements( elements );
-					}
-
-					const billingAddress = billing.billingAddress;
-					const params = {
-						billing_details: {
-							name: `${ billingAddress.first_name } ${ billingAddress.last_name }`.trim(),
-							email: billingAddress.email,
-							phone: billingAddress.phone || null, // Phone is optional, but an empty string is not allowed by Stripe.
-							address: {
-								city: billingAddress.city,
-								country: billingAddress.country,
-								line1: billingAddress.address_1,
-								line2: billingAddress.address_2,
-								postal_code: billingAddress.postcode,
-								state: billingAddress.state,
-							},
-						},
-					};
-					const paymentMethodData = isBlikSelected
-						? {
-								billing_details: params.billing_details,
-								blik: {},
-								type: selectedPaymentMethodType,
-						  }
-						: { elements, params };
-					const paymentMethodObject = await api
-						.getStripe()
-						.createPaymentMethod( paymentMethodData );
-
-					if ( paymentMethodObject.error ) {
-						return {
-							type: 'error',
-							message: paymentMethodObject.error.message,
-						};
-					}
-
-					const dynamicPaymentData = isBlikSelected
-						? {
-								'wc-stripe-blik-code': document?.querySelector(
-									'#wc-stripe-blik-code'
-								)?.value,
-						  }
-						: {};
-
-					return {
-						type: 'success',
-						meta: {
-							paymentMethodData: {
-								...dynamicPaymentData,
-								payment_method: upeMethods[ paymentMethodId ],
-								wc_payment_intent_id: paymentIntentId ?? '',
-								'wc-stripe-is-deferred-intent': true,
-								'wc-stripe-payment-method':
-									paymentMethodObject.paymentMethod.id,
-								save_payment_method: shouldSavePayment
-									? 'yes'
-									: 'no',
-								// The billing information here is relevant to properly create the Stripe Customer object.
-								billing_email: billingAddress.email,
-								billing_first_name: billingAddress.first_name,
-								billing_last_name: billingAddress.last_name,
-								billing_address_1: billingAddress.address_1,
-								billing_address_2: billingAddress.address_2,
-								billing_city: billingAddress.city,
-								billing_state: billingAddress.state,
-								billing_postcode: billingAddress.postcode,
-								billing_country: billingAddress.country,
-							},
-						},
-					};
-				}
-				return handlePaymentProcessing();
-			} ),
-		[
-			activePaymentMethod,
-			api,
-			elements,
-			fingerprint,
-			gatewayConfig,
-			paymentMethodId,
-			paymentMethodsConfig,
-			shouldSavePayment,
-			upeMethods,
-			errorMessage,
-			onPaymentSetup,
-			isPaymentElementComplete,
-			billing.billingAddress,
-			paymentIntentId,
-			selectedPaymentMethodType,
-			isBlikSelected,
-		]
-	);
 
 	useEffect( () => {
 		// Show the Cash App limit notice if the payment method is selected and the cart amount is higher than 2000 USD.
@@ -302,7 +139,23 @@ const PaymentProcessor = ( {
 		}
 	}, [ selectedPaymentMethodType, elements ] );
 
-	usePaymentCompleteHandler(
+	usePaymentSetupHandler(
+		activePaymentMethod,
+		api,
+		billing.billingAddress,
+		elements,
+		errorMessage,
+		hasLoadErrorRef,
+		isPaymentElementComplete,
+		onPaymentSetup,
+		paymentIntentId,
+		paymentMethodId,
+		selectedPaymentMethodType,
+		shouldSavePayment,
+		upeMethods
+	);
+
+	useCheckoutSuccessHandler(
 		api,
 		stripe,
 		elements,
@@ -311,7 +164,7 @@ const PaymentProcessor = ( {
 		shouldSavePayment
 	);
 
-	usePaymentFailHandler(
+	useCheckoutFailHandler(
 		api,
 		stripe,
 		elements,
@@ -346,7 +199,7 @@ const PaymentProcessor = ( {
 					} }
 				/>
 			) }
-			{ isBlikSelected ? (
+			{ selectedPaymentMethodType === PAYMENT_METHOD_BLIK ? (
 				<BlikCodeElement />
 			) : (
 				<PaymentElement
